+++ runtime/com/redhat/ceylon/compiler/java/runtime/model/RuntimeModelLoader.java
@@ -66,6 +66,10 @@ public class RuntimeModelLoader extends ReflectionModelLoader {
         return jars.getContents(path);
     }
 
+    public byte[] getContents(Module module, String path) {
+        return jars.getContents(module, path);
+    }
+    
     @Override
     protected Class<?> loadClass(Module module, String name) {
         ClassLoader classLoader = classLoaders.get(module);
